AP_ADRC: Fix compile errors, rename debug info to pid_info
